Sanitize VBscript protocol (#27)
Introducing `vbscript` protocol sanitization.

Although VB script has been discontinued by Microsoft, it was still
shipping with IE11 until a couple of years ago. Because the disabling
of VBscript in users OSes depends on an OS update, there could still be
many clients out there using browsers capable of executing VBscript
code. On the other hand, the cost of implementing sanitization of the
protocol in the scope of this library is low.